functions/triggerRiskReviewAPI/__init__.py
import logging
import json
import os
import azure.functions as func
from datetime import datetime

# Function to simulate triggering a risk review with an external API
def main(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request to trigger risk review API.')

    try:
        # Parse request body
        req_body = req.get_json()
        task_id = req_body.get('taskId')
        client_name = req_body.get('clientName')
        risk_factors = req_body.get('riskFactors', [])
        
        # Validate input
        if not task_id or not client_name:
            return func.HttpResponse(
                json.dumps({"error": "Missing required parameters: taskId or clientName"}),
                status_code=400,
                mimetype="application/json"
            )
        
        # In a real implementation, this would call an actual external API
        # For the prototype, we'll simulate the API call with a mock response
        
        # Generate a mock tracking ID
        tracking_id = f"RISK-{task_id}-{datetime.now().strftime('%Y%m%d%H%M%S')}"
        
        # Mock external API URL from environment variable
        mock_api_url = os.environ.get('MOCK_RISK_REVIEW_API_URL', 'https://example.com/risk-review')
        
        # Log the simulated API call
        logging.info(f"Simulating risk review API call to {mock_api_url} for client {client_name}")
        
        # In a real environment, we would make an actual API call here
        # For the prototype, generate a mock successful response
        mock_response = {
            "success": True,
            "message": "Risk review request submitted successfully",
            "trackingId": tracking_id,
            "estimatedCompletionTime": "24 hours",
            "riskCategory": "Standard",
            "escalation": False
        }
        
        # Return a success response with mock data
        return func.HttpResponse(
            json.dumps(mock_response),
            status_code=200,
            mimetype="application/json"
        )
    
    except Exception as e:
        logging.error(f"Error triggering risk review API: {str(e)}")
        return func.HttpResponse(
            json.dumps({"error": f"Failed to trigger risk review: {str(e)}"}),
            status_code=500,
            mimetype="application/json"
        )
