functions/requirements.txt
# Azure Functions
azure-functions>=1.13.0

# Azure Storage
azure-storage-blob>=12.15.0
azure-identity>=1.12.0

# Document Processing
python-docx>=0.8.11
docxtpl>=0.16.6
docx2pdf>=0.1.8
PyPDF2>=3.0.0

# Google Drive API (optional for document retrieval)
google-api-python-client>=2.80.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Utilities
requests>=2.28.2
aiohttp>=3.8.4
python-dateutil>=2.8.2
